backend/requirements.txt
fastapi
uvicorn
databases
asyncpg
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-dotenv
geoalchemy2
requests
httpx
python-telegram-bot
slack-sdk
python-jose
passlib[bcrypt]
shapely 
pytest